pre1.2.3.5 现在一条消息只能被其发送者删除了

+++ src/main/webapp/details.jsp
@@ -97,8 +97,10 @@
                     <div class="notification-body">
                         <p class="notification-content"><%= notification.getContent() %></p>
                     </div>
-                    <button class="button-back" onclick="modifyNotification()">Modify</button>
-                    <button class="button-back" onclick="deleteNotification()">Delete</button>
+                    <% if (notification.getPublisherID() == userID) { %>
+                        <button class="button-back" onclick="modifyNotification()">Modify</button>
+                        <button class="button-back" onclick="deleteNotification()">Delete</button>
+                    <% } %>
                     <button class="button-back" onclick="goBack()">Go Back</button>
 
                     <%
